Move notification logic to component
