+++ src/ggrc-client/js/components/page-header/page-header.js
@@ -8,6 +8,7 @@ import '../tasks-counter/tasks-counter';
 import '../tooltip-content/tooltip-content';
 import '../feedback-link/feedback-link';
 import '../release-notes-menu-item/release-notes-menu-item';
+import './notifications-menu-item';
 import logo from '../../../images/ggrc-logo.svg';
 import oneColorLogo from '../../../images/ggrc-one-color.svg';
 import {
