+++ src/ggrc-client/js/components/page-header/page-header.mustache
@@ -140,16 +140,8 @@
         </li>
       {{/if}}
       <li class="notify-wrap">
-        <span>
-          <i class="fa fa-clock-o"></i>
-          Notifications
-        </span>
-        <div class="inner-list">
-          <label>
-            <input type="checkbox" name="notifications" value="Email_Digest">
-            Daily email digest
-          </label>
-        </div>
+        <notifications-menu-item>
+        </notifications-menu-item>
       </li>
       <li>
         <a class="nav-logout" href="/logout" tabindex="-1">
